Require Laravel 6.0+ and PHP 7.2+, drop support for Laravel 5.x and PHP 7.0/7.1
Per my comments in this PR (https://github.com/OveD-php/typed-collections/pull/3):

My suggestion would be to bump the versioning of your Laravel packages to
align with Laravel versioning so that the existing versions (0.0.7, 2.0.2)
remain available for Laravel 5.x users and a new version (e.g. 6.0.0) is
available for Laravel 6.x users.